+++ drools-model/drools-mvel-compiler/src/main/java/org/drools/mvelcompiler/MvelCompiler.java
@@ -3,10 +3,10 @@ package org.drools.mvelcompiler;
 import com.github.javaparser.ast.expr.Expression;
 import org.drools.mvelcompiler.context.MvelCompilerContext;
 import org.drools.mvelcompiler.phase1.StringToExpressionPhase;
-import org.drools.mvelcompiler.phase3.FlattenExpressionPhase;
-import org.drools.mvelcompiler.phase3.FlattenedExpressionResult;
-import org.drools.mvelcompiler.phase2.FirstChildProcessPhase;
-import org.drools.mvelcompiler.phase2.FirstChildProcessResult;
+import org.drools.mvelcompiler.phase2.FlattenExpressionPhase;
+import org.drools.mvelcompiler.phase2.FlattenedExpressionResult;
+import org.drools.mvelcompiler.phase3.FirstChildProcessPhase;
+import org.drools.mvelcompiler.phase3.FirstChildProcessResult;
 import org.drools.mvelcompiler.phase4.ExpressionWithTypePhase;
 import org.drools.mvelcompiler.phase4.TypedExpressions;
 import org.drools.mvelcompiler.phase5.ToJavaExpressionPhase;
