+++ drools-model/drools-mvel-compiler/src/main/java/org/drools/mvelcompiler/phase3/FirstChildProcessResult.java
@@ -3,18 +3,19 @@ package org.drools.mvelcompiler.phase3;
 import java.util.List;
 
 import com.github.javaparser.ast.Node;
+import org.drools.mvelcompiler.phase4.TypedExpression;
 
-public class FlattenedExpressionResult {
+public class FirstChildProcessResult {
 
-    private Node firstNode;
+    private TypedExpression firstNode;
     private List<Node> otherNodes;
 
-    public FlattenedExpressionResult(Node firstNode, List<Node> otherNodes) {
+    public FirstChildProcessResult(TypedExpression firstNode, List<Node> otherNodes) {
         this.firstNode = firstNode;
         this.otherNodes = otherNodes;
     }
 
-    public Node getFirstNode() {
+    public TypedExpression getFirstNode() {
         return firstNode;
     }
